gh-62: Normalized sources and groups
Instead of storing categorical values like source, groups and group
values, created a dedicated table to store the values there. In the main
`mart_ogd_api`, only the IDs of these 3 tables are present and get
joined during API call.

Additionally modified column types (SMALLINT) and indexes to increase
the speed of the api call. Now, the bottleneck is the transformation
from from the DB response into json, xlsx, csv or parquet inside pandas.
Will try out polars/geopolars as soon as they are both ready.

This approach now contains Python assets in between of DBT assets. So
the orchestrator (Dagster) needs to run the jobs.

+++ models/models/marts/dim_bezirk.sql
@@ -41,9 +41,9 @@ with src as (
 select 
     snapshot_code
     , snapshot_date
-    , bezirk_bfs_id
+    , bezirk_bfs_id::SMALLINT
     , bezirk_name
-    , kanton_bfs_id
+    , kanton_bfs_id::SMALLINT
     , geometry
     , EXTRACT(YEAR FROM src.snapshot_date) as snapshot_year
     , ST_Transform(src.geometry, 4326) as geom_border
